% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensirene.R
\name{get_company}
\alias{get_company}
\title{Get company}
\usage{
get_company(siret, endpoint = "http://sirene.entreprise.api.gouv.fr")
}
\arguments{
\item{siret}{a valid siret number}

\item{endpoint}{API endpoint}
}
\value{
a list
}
\description{
Get company
}
\examples{

get_company(siret = "79101208100018")

}
